Add files via upload
Here's a Python function that generates all possible permutations of a given array of distinct integers
This function uses backtracking to generate all permutations of the input array nums. It maintains a path variable to keep track of the current permutation being constructed and a used array to keep track of which elements have been used in the current permutation. The backtrack function recursively explores all possible permutations, adding them to the permutations list when a full permutation is found.